Add layout type support
