+++ Stepic/Sources/Modules/ExploreSubmodules/CatalogBlocksSubmodules/SimpleCourseList/Views/Default/Cells/SimpleCourseListDefaultCollectionViewCell.swift
