+++ Stepic/Sources/Modules/ExploreSubmodules/CatalogBlocksSubmodules/SimpleCourseList/Views/Default/SimpleCourseListCollectionViewDataSource.swift
