+++ Stepic/Sources/Modules/ExploreSubmodules/CatalogBlocksSubmodules/SimpleCourseList/Views/Grid/GridSimpleCourseListView.swift
@@ -1,0 +1,35 @@
+import SnapKit
+import UIKit
+
+extension GridSimpleCourseListView {
+    struct Appearance {}
+}
+
+final class GridSimpleCourseListView: UIView {
+    let appearance: Appearance
+
+    init(
+        frame: CGRect = .zero,
+        appearance: Appearance = Appearance()
+    ) {
+        self.appearance = appearance
+        super.init(frame: frame)
+
+        self.setupView()
+        self.addSubviews()
+        self.makeConstraints()
+    }
+
+    @available(*, unavailable)
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
+extension GridSimpleCourseListView: ProgrammaticallyInitializableViewProtocol {
+    func setupView() {}
+
+    func addSubviews() {}
+
+    func makeConstraints() {}
+}
